requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.6
dataclasses-json==0.6.7
ddgs==9.5.2
distro==1.9.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
grandalf==0.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.4.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lxml==6.0.0
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.0.2
openai==1.99.7
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
primp==0.15.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
